requirements.txt
grpcio==1.49.0
grpcio-tools==1.48.1
matplotlib==3.3.4
numpy==1.22.4
opencv_python_headless==4.5.5.64
pandas==1.2.4
pavi==0.0.1
Pillow==9.2.0
sympy==1.8
protobuf==3.19.4
scikit_learn==1.1.2
scipy==1.7.3
seaborn==0.11.1
statsmodels==0.12.2
sympy==1.8
terminaltables==3.1.10
timm==0.6.7